+++ pyrai/dispatcher/api_calls/fleet/__init__.py
@@ -2,7 +2,20 @@ from .make_vehicle_online import make_vehicle_online
 from .make_vehicle_offline import  make_vehicle_offline
 from .update_vehicle import update_vehicle
 from .remove_vehicle import remove_vehicle
+from .get_vehicle_info import get_vehicle_info
+from .set_params import set_params
+from .add_request import add_request
+from .cancel_request import cancel_request
+from .get_request import get_request
+from .get_assignments import get_assignments
 __all__ = ["make_vehicle_online", 
 "make_vehicle_offline",
 "update_vehicle",
-"remove_vehicle"]
+"remove_vehicle",
+"get_vehicle_info",
+"set_params",
+"add_request",
+"cancel_request",
+"get_request",
+"get_assignments"
+]
