+++ pyrai/dispatcher/structures/__init__.py
@@ -6,6 +6,7 @@ from .fleet_params import FleetParams
 from .event import Event
 from .location import Location
 from .vehicle import Vehicle
+from .fleet import Fleet
 from .metrics import Metrics
 from .user_key import UserKey
 from .request import Request
@@ -13,6 +14,7 @@ from .vehicle_event import VehicleEvent
 from .notification import Notification
 from .notification_data import NotificationData
 from .vehicle_assignments import VehicleAssignments
+from .pyrai import Pyrai
 __all__ = ["Endpoints", 
 "StatusResponse", 
 "StatusError", 
@@ -21,11 +23,13 @@ __all__ = ["Endpoints",
 "Event",
 "Location",
 "Vehicle",
+"Fleet",
 "Metrics",
 "UserKey",
 "Request",
 "VehicleEvent",
 "Notification",
 "NotificationData",
-"VehicleAssignments"
+"VehicleAssignments",
+'Pyrai'
 ]
